移除UIButton分类废弃方法 宏文件修改 0.3.0

+++ RdUtils/RdUtils/RondoUtils/RdMacroFile.h
@@ -14,7 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 #define Rd_AnimationDuration    0.26f
 #define Rd_BarButtonTitleFontSize  16
-#define Rd_IGNORE               CGFLOAT_MAX
+#define Rd_IGNORE               (CGFloat)CGFLOAT_MAX
 
 #define Rd_WeakSelf(type)        __weak typeof(type) weak##type = type;
 #define Rd_StrongSelf(type)      __strong typeof(type) type = weak##type;
